requirements.txt
flask==3.1.0
requests==2.32.3
feedparser==6.0.11
google-genai==1.10.0
python-dotenv==1.1.0
